New attack animation
NOTE: Maya binary files (.mb) are dependent upon Maya being on a user's machine. FBX files are cumbersome to get to work atm, but we should look keep this in mind when fully animating Niko.

Avatar Mask applied to upper body to allow character to move/jump and still attack.