Implemented Object Interaction
Player can now interact with objects as well as NPCs with the same button!

To make an object interactable by touch:
- Make a script for the object that inherits Ninjacat.Utility.IInteract instead of MonoBehaviour.
- Implement the abstract method handleInteraction(GameObject obj)
- If necessary, override virtual methods confirmInteraction(GameObject obj) and endInteraction(GameObject obj)

Added some new targeting utility functions:
-conicoidLOS
-wedgeLOS

actOnLayer() can now be called with any of the 3 existing targeting functions.
- It defaults to wedgeLOS

Fixed an issue with controls that are held down occassionally releasing for a frame.